+++ R/compute_nsize.R
@@ -1,6 +1,8 @@
 #' Check for N column if not present and user wants, impute N based on user's
 #' sample size. **NOTE** this will be the same value for each SNP which is not
-#' necessarily correct and may cause issues down the line.
+#' necessarily correct and may cause issues down the line. N can also be 
+#' inputted with "ldsc", "sum", "giant" or "metal" by passing one or 
+#' multiple of these.
 #'
 #' @return \code{list("sumstats_dt"=sumstats_dt)}
 #'
@@ -27,25 +29,38 @@
 #' @keywords internal
 compute_nsize <- function(sumstats_dt,
                           imputation_ind = FALSE,
-                          compute_n = c("ldsc", "sum"),
+                          compute_n = c("ldsc", "giant", "metal", "sum"),
                           force_new = FALSE) {
     ## Avoid confusing BiocCheck.
-    IMPUTATION_n <- NULL
-
+    IMPUTATION_n = IMPUTATION_Neff = NULL
+    
+    #if you want an Neff column for more than one method need to ensure
+    #you can tell which is which
+    append_method_name <- FALSE
+    if(is.vector(compute_n)){
+        #sum makes an N column not an Neff column 
+        if(length(compute_n[!compute_n=="sum"])>1)
+            append_method_name <- TRUE
+    }
     for (method in compute_n) {
         compute_sample_size(
             sumstats_dt = sumstats_dt,
             method = method,
-            force_new = force_new
+            force_new = force_new,
+            append_method_name = append_method_name
         )
     }
     # if user wants information, give SNPs where Z-score calculated
     ## Evaluate the conditions under which N would have been calculate by
     ## compute_sample_size()
-    if (imputation_ind &
-        is.numeric(compute_n) &
+    if (imputation_ind &&
+        is.numeric(compute_n) && compute_n!=0L &&
         (!"N" %in% names(sumstats_dt))) {
         sumstats_dt[, IMPUTATION_n := TRUE]
+    }
+    if (imputation_ind &&
+        is.character(compute_n)) {
+        sumstats_dt[, IMPUTATION_Neff := TRUE]
     }
     return(list("sumstats_dt" = sumstats_dt))
 }