+++ tests/testthat/test-check_imputation_cols.R
@@ -45,7 +45,24 @@ test_that("Check that imputation columns added correctly", {
               }
         }
         expect_equal(have_value, TRUE)
+        
+        #check other compute_n values
+        eduAttainOkbay <- data.table::fread(pth)
+        eduAttainOkbay[,N_CON:=100]
+        eduAttainOkbay[,N_CAS:=120]
+        # write to temp dir
+        file <- tempfile()
+        data.table::fwrite(eduAttainOkbay, file)
+        methods <- c("ldsc", "sum", "giant", "metal")
+        reformatted <- MungeSumstats::format_sumstats(file,
+                                                      ref_genome = "GRCh37",
+                                                      compute_n = methods)
+        res <- data.table::fread(reformatted)
+        expect_equal(all(paste0("Neff_",c("ldsc","giant","metal")) %in% 
+                             colnames(res)),TRUE)
+        
     } else {
+        expect_equal(is_32bit_windows, TRUE)
         expect_equal(is_32bit_windows, TRUE)
     }
 })